ProveIt_Project/Code/proto_v0.19/proveit/homie/models/rasa/actions/requirements.txt
python-dotenv==1.0.1
Requests==2.31.0
sentence_transformers==2.5.1
torch==2.2.1
